Changed references of p327 to p286
Not sure why there's a reference in the install instructions to 1.9.3-p327, a non-existent (as far as I can tell) version. I know rbenv isn't written for those who simply copy and paste their way through app development, but putting in a seemingly broken value without warning that it's a placeholder seems unnecessary. When p327 didn't work, I tried the version of Ruby that I have now (1.9.3-p258) and got the same error. At that point, I didn't know if rbenv + ruby-build is limited to very specific versions or if either rbenv and/or ruby-build wasn't properly installed on my machine.