URL hack to fix recognizing that Topaz archive is bz2
On OS X and some other platforms it doesn't matter, as `tar`
automatically detects the compression method, but on platforms that
don't it's important to declare the archive format. For ordinary
tarballs the extension provides this info, but since Topaz archive
doesn't have an extension in the URL, we add a fake one via a dummy
query string parameter.

Fixes #357