Add option to not install existing versions
The -f option allows for forcing installation of versions that already
exist in the environment. If that option isn't used, then an interactive
prompt is created which causes problems for non-interactive execution of
the command. One could use -f for non-interactive executions, but it
causes a significant slowdown in runtime if you don't want to reinstall
that version.

This change allows for a -n option to not install versions that already
exist, while still exiting with a happy status code (for cases where
scripts may be kicking off the rbenv install command where `set -e` has
been run).