Add support for $RUBY_BUILD_DEFINITIONS.
$RUBY_BUILD_DEFINITIONS is a colon-separated series of paths for
definitions. Each path is the file system location of a directory
in which definitions may be found. If the definition being installed
is not distributed with ruby-build, the program will iterate through
each of these paths and look within them (just the base directory)
for the definition being installed. If it is found, the full path to
the definition will be set as the $DEFINITION_PATH and will be installed.

Resolves https://github.com/sstephenson/ruby-build/issues/608.