Avoid double shebangs when fixing Rubinius gem binstubs
This is to fix the error when installing new gems that have executables
which match existing binstubs in the Rubinius bin directory:

    "bundle" from bundler conflicts with PREFIX/gems/bin/bundle

RubyGems is supposed to override the binstub if it detects that the
previous one was also generated by RubyGems for the gem of the same
name, but its detection mechanism gets thrown off by having a double
shebang as a result of our binstubs fixing process.

https://github.com/rubygems/rubygems/blob/v2.2.2/lib/rubygems/installer.rb#L149-L154

This avoids generating binstubs with a double shebang.