Allow for space-separated LDFLAGS
Prevented errors caused by a space between `-L` and the path by not creating a directory if the arg is `-L`

The base case of the switch statement will be called with
either `-L/path/` or `/path/`. `mkdir` will then be called
with that argument after the `-L` has been removed.