Restrict find command to subdirectories
Fixes #902

The find command executed by `fix_directory_permissions` is too
inclusive.  It searches everything in the `$PREFIX_DIR`, which includes
directories that ruby-build does not care about, such as `lost+found`.
By restricting the find command to the `$PREFIX_DIR/bin` directory and
the `$PREFIX_DIR/share` directory, ruby-build will only change
permissions on the directories it touches, ignoring anything else that
might be in the `$PREFIX_DIR`, which could be anything.