bin/ruby-build: support DESTDIR setting
We can't just use MAKE_INSTALL_OPTS because DESTDIR modifies the
location of the installed ruby binary, so anything that used the old
RUBY_BIN also failed - it needed to be modified as well.

Instead introduce a new variable that edits both RUBY_BIN and
MAKE_INSTALL_OPTS if set.

Also edit LDFLAGS so we can find libruby.so in the DESTDIR for
post-install checks.

Fixes #1415.
Updates #413.