Don't delete jruby.sh
I'm not sure why this file was being deleted but this code dates way back! Unfortunately deleting this interferes with a new version of the native JRuby launcher we want to release which is dependent on the shell script being present. This PR removes jruby.sh from the `remove_windows_files` function so it can be used by the new launcher.

See https://github.com/jruby/jruby-launcher/pull/48#issuecomment-2772272290